wyp_backend/requirements.txt
django
psycopg2-binary
djangorestframework
Pillow
django-filter